Menu

+++ modules/filesystem/src/components/FilesBehavior.php
@@ -182,18 +182,20 @@ class FilesBehavior extends Behavior implements ModelBehaviorInterface
     protected function deleteRelatedFiles()
     {
         $this->normalizeAttributes();
+        $filesystemComponent = $this->getFilesystemComponent();
         foreach ($this->attributes as $attribute => $storageConfig) {
             $delete = $storageConfig->deleteOnUnlink || $this->owner->getRelationSetting($storageConfig->relation, 'deleteOnUnlink');
-            $filesystemComponent = $this->getFilesystemComponent();
-            /** @var FileInterface[] $files */
-            $files = $this->owner->getRelation($storageConfig->relation)->all();
-            foreach ($files as $file) {
-                $filesystemComponent->deleteFile($file);
+            if ($delete) {
+                /** @var FileInterface[] $files */
+                $files = $this->owner->getRelation($storageConfig->relation)->all();
+                foreach ($files as $file) {
+                    $filesystemComponent->deleteFile($file);
+                }
+                if (isset($storageConfig->events['afterDeleteAll']) && $storageConfig->events['afterDeleteAll'] instanceof \Closure) {
+                    call_user_func($storageConfig->events['afterDeleteAll'], $storageConfig, $filesystemComponent);
+                }
             }
             $this->owner->unlinkAll($storageConfig->relation, $delete);
-            if (isset($storageConfig->events['afterDeleteAll']) && $storageConfig->events['afterDeleteAll'] instanceof \Closure) {
-                call_user_func($storageConfig->events['afterDeleteAll'], $storageConfig, $filesystemComponent);
-            }
         }
     }
 
@@ -225,26 +227,25 @@ class FilesBehavior extends Behavior implements ModelBehaviorInterface
     protected function linkRelatedFiles()
     {
         if ($this->_relatedFiles) {
+            $filesystemComponent = $this->getFilesystemComponent();
             foreach ($this->attributes as $attribute => $storageConfig) {
                 if (array_key_exists($attribute, $this->_relatedFiles)) {
-
-//                    $extraColumns = $storageConfig->extraColumns || $this->owner->getRelationSetting($storageConfig->relation, 'extraColumns');
-//                    $extraColumns = $extraColumns ?: [];
-//                    $relation = $this->owner->getRelation($storageConfig->relation);
-//                    if ($extraColumns && $relation->via === null) {
-//                        foreach ($this->_relatedFiles[$attribute]['models'] as $model) {
-//                            foreach ($extraColumns as $attr => $value) {
-//                                $model->$attr = $value;
-//                            }
-//                        }
-//                    }
-//                    foreach ($this->_relatedFiles[$attribute]['models'] as $model) {
-//                        $this->owner->link($storageConfig->relation, $model, $extraColumns);
-//                    }
-
-                    $this->owner->{$storageConfig->relation} = $this->_relatedFiles[$attribute]['models'];
-
+                    $models = $this->_relatedFiles[$attribute]['models'];
+                    $this->owner->{$storageConfig->relation} = $models;
                     unset($this->_relatedFiles[$attribute]);
+                    if ($models === null) {
+                        $delete = $storageConfig->deleteOnUnlink || $this->owner->getRelationSetting($storageConfig->relation, 'deleteOnUnlink');
+                        if ($delete) {
+                            /** @var FileInterface[] $files */
+                            $files = $this->owner->getRelation($storageConfig->relation)->all();
+                            foreach ($files as $file) {
+                                $filesystemComponent->deleteFile($file);
+                            }
+                            if (isset($storageConfig->events['afterDeleteAll']) && $storageConfig->events['afterDeleteAll'] instanceof \Closure) {
+                                call_user_func($storageConfig->events['afterDeleteAll'], $storageConfig, $filesystemComponent);
+                            }
+                        }
+                    }
                 }
             }
             $this->owner->save(false);
